+++ pyavrocd/deviceinfo/harvest.py
@@ -315,6 +315,22 @@ def capture_signature_from_property_groups_element(element):
             signature[2] = int(i.attrib['value'], 16)
     return signature
 
+def capture_eesave_from_register_group(element):
+    """
+    Capture the fuse address of the EESAVE bit from a register-group element
+
+    :param element: element with tag='register-group'
+    :type element: xml.etree.ElementTree.Element instance
+    :return: fuse address where EESAVE is located
+    :rtype: int
+    """
+    for reg in element.findall('register'):
+        off = reg.attrib['offset']
+        for bf in reg.findall('bitfield'):
+            if bf.attrib['name'].lower() == 'eesave':
+                return off
+    return None
+
 def capture_bootrst_from_register_group(element):
     """
     Capture the fuse address of the BOOTRST bit from a register-group element
@@ -488,6 +504,8 @@ def harvest_from_file(filename):
     ocden_mask = None
     bootrst_base = None
     bootrst_mask = None
+    eesave_base = None
+    eesave_mask = None
     buf_per_page = None
 
     memories = {}
@@ -552,6 +570,9 @@ def harvest_from_file(filename):
                     ocden_mask = elem.attrib['mask']
                 if elem.attrib['name'].lower() == 'bootrst':
                     bootrst_mask = elem.attrib['mask']
+                if elem.attrib['name'].lower() == 'eesave':
+                    eesave_mask = elem.attrib['mask']
+
             if elem.tag == 'register-group':
                 if capture_bootrst_from_register_group(elem) is not None:
                     bootrst_base = capture_bootrst_from_register_group(elem)
@@ -559,6 +580,8 @@ def harvest_from_file(filename):
                     dwen_base = capture_dwen_from_register_group(elem)
                 if capture_ocden_from_register_group(elem) is not None:
                     ocden_base = capture_ocden_from_register_group(elem)
+                if capture_eesave_from_register_group(elem) is not None:
+                    eesave_base = capture_eesave_from_register_group(elem)
 
 
     extra_fields += capture_field('address_size', determine_address_size(progmem_offset))
@@ -607,6 +630,10 @@ def harvest_from_file(filename):
         extra_fields += "    'bootrst_base' : " + "%s" % bootrst_base + ",\n"
     if bootrst_mask:
         extra_fields += "    'bootrst_mask' : " + "%s" % bootrst_mask + ",\n"
+    if eesave_base:
+        extra_fields += "    'eesave_base' : " + "%s" % eesave_base + ",\n"
+    if eesave_mask:
+        extra_fields += "    'eesave_mask' : " + "%s" % eesave_mask + ",\n"
     if buf_per_page:
         extra_fields += "    'buffers_per_flash_page' : " + "%s" % buf_per_page + ",\n"
     if masked: